cleaning up 1559 gas input
