+++ client/src/middlewares/submitLevelInstance.js
@@ -1,6 +1,6 @@
 import * as actions from '../actions';
 import { loadTranslations } from '../utils/translations'
-import * as constants from "../constants";
+import { getGasFeeDetails } from '../utils/ethutil'
 
 let language = localStorage.getItem('lang')
 let strings = loadTranslations(language)
@@ -19,13 +19,13 @@ const submitLevelInstance = store => next => async action => {
   ) return next(action)
 
   console.asyncInfo(`@good ${strings.submitLevelMessage}`)
-  const gasDetails = await getGasDetails(state.network)
+  const gasFeeDetails = await getGasFeeDetails(state.network)
   let completed = await submitLevelInstanceUtil(
     state.contracts.ethernaut,
     action.level.deployedAddress,
     state.contracts.levels[action.level.deployedAddress].address,
     state.player.address,
-    gasDetails
+    gasFeeDetails
   )
   if(completed) {
     console.victory(`@good ${strings.wellDoneMessage}, ${strings.completedLevelMessage}`)
@@ -40,9 +40,9 @@ const submitLevelInstance = store => next => async action => {
 
 export default submitLevelInstance
 
-async function submitLevelInstanceUtil(ethernaut, levelAddress, instanceAddress, player, gasDetails) {
+async function submitLevelInstanceUtil(ethernaut, levelAddress, instanceAddress, player, gasFeeDetails) {
   try {
-    const data = { from: player, ...gasDetails }
+    const data = { from: player, ...gasFeeDetails }
     const tx = await ethernaut.submitLevelInstance(instanceAddress, data);
     if (tx.logs.length === 0) return false
     else {
@@ -62,23 +62,3 @@ async function submitLevelInstanceUtil(ethernaut, levelAddress, instanceAddress,
     return false
   }
 }
-
-const getGasDetails = async (network) => {
-  if (
-    network.networkId.toString() === constants.NETWORKS.MUMBAI.id ||
-    network.networkId.toString() === constants.NETWORKS.SEPOLIA.id ||
-    network.networkId.toString() === constants.NETWORKS.GOERLI.id
-  ) {
-    const maxPriorityFeePerGas = network.web3.utils.toWei('2.5', 'gwei');
-    const block = await network.web3.eth.getBlock('latest')
-    const blockBaseFee = block.baseFeePerGas ? block.baseFeePerGas : 1;
-    return {
-      maxPriorityFeePerGas,
-      maxFeePerGas: 2 * Number(blockBaseFee) + Number(maxPriorityFeePerGas)
-    }
-  } else { 
-    return {
-      gasPrice: network.gasPrice
-    }
-  }
-}
